refactor(worksheets): share the inline rename between surfaces
The rename state machine and its input lived inside the explorer, so the
tab strip could not reuse them. Move the hook to
hooks/use-worksheet-rename and the field to WorksheetNameInput, leaving
each caller to supply its own wrapper and sizing.

The field also gains an aria-label — it was an unlabelled textbox.
